Added feed.repo
