+++ src/modules/feedParser/types/feed.types.ts
@@ -1,0 +1,12 @@
+export type NormalizedFeedItem = {
+    title?: string;
+    link?: string;
+    isoDate?: string;
+    pubDate?: string;
+    description?: string;
+};
+
+export type NormalizedFeed = {
+    sourceUrl: string;
+    items: NormalizedFeedItem[];
+};
